backend/requirements.txt
fastapi
uvicorn
python-dotenv
langchain
langchain-community
langchain-openai
openai
faiss-cpu
chromadb
pyairtable
pydantic